Add configuration and logging entities; refactor signal handling and evaluation logic

+++ Sigmentum/Models/Candle.cs
@@ -3,9 +3,9 @@
 public class Candle
 {
     public DateTime Time { get; set; }
-    public double Open { get; set; }
-    public double High { get; set; }
-    public double Low { get; set; }
-    public double Close { get; init; }
-    public double Volume { get; set; }
+    public decimal Open { get; set; }
+    public decimal High { get; set; }
+    public decimal Low { get; set; }
+    public decimal Close { get; init; }
+    public decimal Volume { get; set; }
 }
