+++ Sigmentum/Services/Indicators.cs
@@ -2,10 +2,10 @@
 
 public static class Indicators
 {
-    public static List<double> CalculateRsi(List<double> prices, int period)
+    public static List<decimal> CalculateRsi(List<decimal> prices, int period)
     {
-        var result = new List<double>();
-        double gain = 0, loss = 0;
+        var result = new List<decimal>();
+        decimal gain = 0, loss = 0;
         for (var i = 1; i <= period; i++)
         {
             var diff = prices[i] - prices[i - 1];
@@ -31,16 +31,16 @@ public static class Indicators
                 loss = (loss * (period - 1) - diff) / period;
             }
 
-            result.Add(100 - 100 / (1 + gain / (loss == 0 ? 0.01 : loss)));
+            result.Add(100 - 100 / (1 + gain / (loss == 0 ? 0.01m : loss)));
         }
 
-        return Enumerable.Repeat(50.0, prices.Count - result.Count).Concat(result).ToList();
+        return Enumerable.Repeat(50.0m, prices.Count - result.Count).Concat(result).ToList();
     }
 
-    public static List<double> CalculateEma(List<double> prices, int period)
+    public static List<decimal> CalculateEma(List<decimal> prices, int period)
     {
-        var ema = new List<double>();
-        var multiplier = 2.0 / (period + 1);
+        var ema = new List<decimal>();
+        var multiplier = 2.0m / (period + 1);
         ema.Add(prices.Take(period).Average());
 
         for (var i = period; i < prices.Count; i++)
